Implement project layout foundation
Split implementation into a frontend reference application and core
library, which allows for straightforward unit testing and eases the
potential work needed to allow for the use of the core algorithim in
other projects.

Make the classes involved in the STAR calculation as generic as
possible, with specifics related to the target use case (PFCC voting)
segerated to the "reference" portion of the library, used by the
reference application.

Overall idea is to add a logger, then have the reference application
simply act as a command line interface and data presenter, using the lib
to load the voting data from the reference format and then calculate the
results. The lib's main routine should be implemented as a QObject
derived class that emits signals to provide calculation details, but
otherwise simply returns a results object (report?). The results object
may potentially include a few secondary high level details that describe
how the results were determined.

The front end can then use the election data is already has access to,
along with the results in order to present the categories, nominees, and
results of each election.