Change Condorcet extra tie break method to be more proper
rankByHeadToHeadWins() now performs head-to-heads of each nominee up for
consideration (i.e. part of its argument list) against all other
nominees, instead of just those that are part of the tie.

Although the examples of this method don't show it being used as an
extra tiebreaker (even though it says you can, and recommends you use it
as such), it's clear that all head-to-heads matter, even if only a sub-
set of the participants are under consideration when counting the wins.