Use CMake --install switch instead of 'install' target
CMake's built in install scripts will only install what has been built,
which is desired, instead of when invoking ninja where the install
target causes everything under the selected configuration to be built.

CMAKE_SKIP_INSTALL_ALL_DEPENDENCY is supposed to be able to remedy this,
but initial testing made it seem that it does not work, at least not
without further changes.

This was an issue since the intention is to only build the library in
Debug configuration and not the application.