Actions: Build 'all' target for Release so that tests are included
CMake doesn't seem to have a convenient, built-in way to build all test
targets, either requiring one to build all of them explicitly or to
create a custom target that depends on all tests. Since here we do
actually want all Release targets to be built, it's fine to simply use
'all' in that configuration.

In the future if there is ever a reason to only build a subset of
targets in Release mode then this fundamental issue will still present a
problem.